+++ lib/src/typography/tk_font_family.dart
@@ -1,8 +1,0 @@
-class TKFontFamily {
-  const TKFontFamily._(this.name);
-  final String name;
-
-  static const TKFontFamily dufanthe = TKFontFamily._("Dufanthe");
-  static const TKFontFamily kumbhSans = TKFontFamily._("KumbhSans");
-  static const TKFontFamily inter = TKFontFamily._("Inter");
-}